added untested code for bootloader commands readInfo and readFlash

+++ ClientSocket.h
@@ -30,7 +30,7 @@ class ClientSocket
 	public:
 		ClientSocket(const char* pAddr, short port, int style);
 		virtual ~ClientSocket();
-		virtual int Recv(unsigned char* pBuffer, size_t length, unsigned long timeoutTmms = 0) const = 0;
+		virtual size_t Recv(unsigned char* pBuffer, size_t length, unsigned long timeoutTmms = 0) const = 0;
 		virtual int Send(const unsigned char* frame, size_t length) const = 0;
 };
 
@@ -38,7 +38,7 @@ class TcpSocket : public ClientSocket
 {
 	public:
 		TcpSocket(const char* pAddr, short port);
-		virtual int Recv(unsigned char* pBuffer, size_t length, unsigned long timeoutTmms = 0) const;
+		virtual size_t Recv(unsigned char* pBuffer, size_t length, unsigned long timeoutTmms = 0) const;
 		virtual int Send(const unsigned char* frame, size_t length) const;
 };
 
@@ -46,7 +46,7 @@ class UdpSocket : public ClientSocket
 {
 	public:
 		UdpSocket(const char* pAddr, short port);
-		virtual int Recv(unsigned char* pBuffer, size_t length, unsigned long timeoutTmms = 0) const;
+		virtual size_t Recv(unsigned char* pBuffer, size_t length, unsigned long timeoutTmms = 0) const;
 		virtual int Send(const unsigned char* frame, size_t length) const;
 };
 #endif /* #ifndef _CLIENTSOCKET_H_ */
